requirements.txt
aniso8601==9.0.1
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.7
fastapi==0.95.1
graphene==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pycparser==2.22
pydantic==1.10.18
PyJWT==2.7.0
sniffio==1.3.1
SQLAlchemy==2.0.11
starlette==0.26.1
starlette-graphene3==0.6.0
typing_extensions==4.12.2
uvicorn==0.21.1
